tests: Add build tags to tests
